ajustes nos arquivos de som

+++ Unity/TCP IV Ultimo Prototipo/Assets/Scripts/Buraco/Buraco.cs
@@ -4,7 +4,7 @@ public class Buraco : MonoBehaviour
 {
     public bool EstaOcupado { get; set; }
 
-    public void CriarToupeira(GameObject toupeira, Toupeira toupeiraAtr, int bpm, AudioClip somAoSurgir, AudioClip somPancada, GameObject acertouEfeito)
+    public void CriarToupeira(GameObject toupeira, Toupeira toupeiraAtr, float bpm, AudioClip somAoSurgir, AudioClip somPancada, GameObject acertouEfeito)
     {
         Vector3 pos = this.gameObject.transform.position + new Vector3(0, -2.8f, 0);
         toupeira = Instantiate(toupeira, pos, Quaternion.identity);
